revmc_context

Struct EvmStack

#[repr(C)]
pub struct EvmStack(pub(crate) [MaybeUninit<EvmWord>; 1024]);

EVM context stack.

Tuple Fields
0: [MaybeUninit<EvmWord>; 1024]
Implementations

impl EvmStack

pub const SIZE: usize = 32_768usize

The size of the stack in bytes.

pub const CAPACITY: usize = 1_024usize

The size of the stack in U256 elements.

pub fn new() -> Self

Creates a new EVM stack, allocated on the stack.

Use EvmStack::new_heap to create a stack on the heap.

pub fn new_heap() -> Vec<EvmWord>

Creates a vector that can be used as a stack.

pub fn from_interpreter_stack(stack: &mut Stack) -> (&mut Self, &mut usize)

Creates a stack from the interpreter’s stack. Assumes that the stack is large enough.

pub fn from_vec(vec: &Vec<EvmWord>) -> &Self

Creates a stack from a vector’s buffer.

Panics

Panics if the vector’s capacity is less than the required stack capacity.

pub fn from_mut_vec(vec: &mut Vec<EvmWord>) -> &mut Self

Creates a stack from a mutable vector’s buffer.

The bytecode function will overwrite the internal contents of the vector, and will not set the length. This is simply to have the stack allocated on the heap.

Panics

Panics if the vector’s capacity is less than the required stack capacity.

Examples

use revmc_context::EvmStack;
let mut stack_buf = EvmStack::new_heap();
let stack = EvmStack::from_mut_vec(&mut stack_buf);
assert_eq!(stack.as_slice().len(), EvmStack::CAPACITY);

pub const fn from_slice(slice: &[EvmWord]) -> &Self

Creates a stack from a slice.

Panics

Panics if the slice’s length is less than the required stack capacity.

pub fn from_mut_slice(slice: &mut [EvmWord]) -> &mut Self

Creates a stack from a mutable slice.

Panics

Panics if the slice’s length is less than the required stack capacity.

pub const unsafe fn from_ptr<'a>(ptr: *const EvmWord) -> &'a Self

Creates a stack from a pointer.

Safety

The caller must ensure that the pointer is valid and points to at least EvmStack::SIZE bytes.

pub unsafe fn from_mut_ptr<'a>(ptr: *mut EvmWord) -> &'a mut Self

Creates a stack from a mutable pointer.

Safety

The caller must ensure that the pointer is valid and points to at least EvmStack::SIZE bytes.

pub const fn as_bytes(&self) -> &[u8; 32768]

Returns the stack as a byte array.

pub fn as_bytes_mut(&mut self) -> &mut [u8; 32768]

Returns the stack as a byte array.

pub const fn as_slice(&self) -> &[EvmWord; 1024]

Returns the stack as a slice.

pub fn as_mut_slice(&mut self) -> &mut [EvmWord; 1024]

Returns the stack as a mutable slice.