revmc_context

Struct EvmWord

#[repr(C, align(8))]
pub struct EvmWord(pub(crate) [u8; 32]);

A native-endian 256-bit unsigned integer, aligned to 8 bytes.

This is a transparent wrapper around [U256] on little-endian targets.

Tuple Fields
0: [u8; 32]
Implementations

impl EvmWord

pub const ZERO: Self

The zero value.

pub const fn from_ne_bytes(x: [u8; 32]) -> Self

Creates a new value from native-endian bytes.

pub fn from_be_bytes(x: [u8; 32]) -> Self

Creates a new value from big-endian bytes.

pub fn from_le_bytes(x: [u8; 32]) -> Self

Creates a new value from little-endian bytes.

pub fn from_be(x: Self) -> Self

Converts an integer from big endian to the target’s endianness.

pub fn from_le(x: Self) -> Self

Converts an integer from little endian to the target’s endianness.

pub const fn from_u256(value: U256) -> Self

Converts a [U256].

pub const fn from_u256_ref(value: &U256) -> &Self

Available on little-endian only.

Converts a [U256] reference to a [U256].

pub fn from_u256_mut(value: &mut U256) -> &mut Self

Available on little-endian only.

Converts a [U256] mutable reference to a [U256].

pub fn to_be_bytes(self) -> [u8; 32]

Return the memory representation of this integer as a byte array in big-endian (network) byte order.

pub fn to_le_bytes(self) -> [u8; 32]

Return the memory representation of this integer as a byte array in little-endian byte order.

pub const fn to_ne_bytes(self) -> [u8; 32]

Return the memory representation of this integer as a byte array in native byte order.

pub fn to_be(self) -> Self

Converts self to big endian from the target’s endianness.

pub fn to_le(self) -> Self

Converts self to little endian from the target’s endianness.

pub fn swap_bytes(self) -> Self

Reverses the byte order of the integer.

pub const fn as_u256(&self) -> &U256

Available on little-endian only.

Casts this value to a [U256]. This is a no-op on little-endian systems.

pub fn as_u256_mut(&mut self) -> &mut U256

Available on little-endian only.

Casts this value to a [U256]. This is a no-op on little-endian systems.

pub const fn to_u256(&self) -> U256

Converts this value to a [U256]. This is a simple copy on little-endian systems.

pub const fn into_u256(self) -> U256

Converts this value to a [U256]. This is a no-op on little-endian systems.

pub fn to_address(self) -> Address

Converts this value to an [Address].